refactor: centralize shared widget constants
